requirements.txt
google-generativeai==0.5.4
